requirements.txt
asgiref==3.5.0
black==22.3.0
click==8.1.2
Django==4.0.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
python-dotenv==0.20.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
